# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HORNET VPN API",
    "description": "Enterprise-grade distributed VPN management platform",
    "version": "2.0.0",
    "contact": {
      "name": "HORNET Team",
      "url": "https://hornet.com"
    }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Development" },
    { "url": "https://api.hornet.com", "description": "Production" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "nodeTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-node-token"
      }
    }
  },
  "paths": {
...
    "/api/admin/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Admin login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 8 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/admin/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get current admin profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Admin profile" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/api/nodes/register": {
      "post": {
        "tags": ["Nodes"],
        "summary": "Register a VPS node",
        "security": [{ "nodeTokenAuth": [] }],
        "responses": {
          "201": { "description": "Node registered" }
        }
      }
    },
    "/api/nodes/heartbeat": {
      "post": {
        "tags": ["Nodes"],
        "summary": "Send node heartbeat",
        "security": [{ "nodeTokenAuth": [] }],
        "responses": {
          "200": { "description": "Heartbeat acknowledged" }
        }
      }
    },
...
  }
}